PR#16

+++ scripts/test.sh
@@ -2,34 +2,41 @@
 
 init()
 {
+	exit_code=1
+	vm_exec=export.txt
+
+	vm_release=master
 	output_time=0.0
 	wait_for=2
-	vm_release=master
 
-	test_dir=../tests
-	error_dir=../tests/errors
-	exec_dir=../tests/executable
+	dir_install=./install
+	dir_test=../tests
+	dir_error=../tests/errors
+	dir_exec=../tests/executable
 
-	error_subdir=errors
-	warning_subdir=warnings
-	include_subdir=include
+	subdir_error=errors
+	subdir_warning=warnings
+	subdir_include=include
 
 	while ! [[ -z $1 ]]
 	do
-		case "$1" in
+		case $1 in
 			-h|--help)
 				echo -e "Usage: ./${0##*/} [KEY] ..."
 				echo -e "Description:"
-				echo -e "\tThis script tests all files from \"$test_dir\" directory."
-				echo -e "\tFolder \"$error_dir\" should contain tests with expected error."
-				echo -e "\tExecutable tests should be in \"$exec_dir\" directory."
-				echo -e "\tTo ignore invalid tests output, use \"*/$warning_subdir/*\" subdirectory."
-				echo -e "\tFor tests with expected runtime error, use \"*/$error_subdir/*\" subdirectory."
-				echo -e "\tFor multi-file tests, use \"*/$include_subdir/*\" subdirectory."
+				echo -e "\tThis script tests all files from \"$dir_test\" directory."
+				echo -e "\tFolder \"$dir_error\" should contain tests with expected error."
+				echo -e "\tExecutable tests should be in \"$dir_exec\" directory."
+				echo -e "\tTo ignore invalid tests output, use \"*/$subdir_warning/*\" subdirectory."
+				echo -e "\tFor tests with expected runtime error, use \"*/$subdir_error/*\" subdirectory."
+				echo -e "\tFor multi-file tests, use \"*/$subdir_include/*\" subdirectory."
+				echo -e "\tFailed tests for debug build only will be marked with \"(Debug)\"."
 				echo -e "Keys:"
 				echo -e "\t-h, --help\tTo output help info."
 				echo -e "\t-s, --silence\tFor silence testing."
+				echo -e "\t-f, --fast\tFast testing, Release builds only."
 				echo -e "\t-i, --ignore\tIgnore errors & executing stages."
+				echo -e "\t-r, --remove\tRemove build folder before testing."
 				echo -e "\t-d, --debug\tSwitch on debug tracing."
 				echo -e "\t-v, --virtual\tSet RuC virtual machine release."
 				echo -e "\t-o, --output\tSet output printing time (default = 0.0)."
@@ -39,8 +46,14 @@ init()
 			-s|--silence)
 				silence=$1
 				;;
+			-f|--fast)
+				fast=$1
+				;;
 			-i|--ignore)
 				ignore=$1
+				;;
+			-r|--remove)
+				remove=$1
 				;;
 			-d|--debug)
 				debug=$1
@@ -66,40 +79,101 @@ init()
 	failure=0
 	timeout=0
 
+	if [[ $OSTYPE == "darwin" ]] ; then
+		runner="gtimeout $wait_for"
+	else
+		runner="timeout $wait_for"
+	fi
+
 	log=tmp
+	buf=buf
+}
+
+build_folder()
+{
+	if ! [[ -z $remove ]] ; then
+		rm -rf build
+	fi
+	mkdir -p build && cd build
+
+
+	if [[ -z $fast ]] ; then
+		if [[ $OSTYPE != "msys" ]] ; then
+			CMAKE_BUILD_TYPE=-DCMAKE_BUILD_TYPE=Debug
+		fi
+
+		cmake .. $CMAKE_BUILD_TYPE -DTESTING_EXIT_CODE=$exit_code
+		if ! cmake --build . --config Debug ; then
+			exit 1
+		fi
+
+		if [[ $OSTYPE != "msys" ]] ; then
+			cmake --install . --prefix $dir_install --config Debug
+			rm -rf Debug
+			mv $dir_install/$1 Debug
+		fi
+	fi
+
+	if [[ $OSTYPE != "msys" ]] ; then
+		CMAKE_BUILD_TYPE=-DCMAKE_BUILD_TYPE=Release
+	fi
+
+	cmake .. $CMAKE_BUILD_TYPE -DTESTING_EXIT_CODE=$exit_code
+	if ! cmake --build . --config Release ; then
+		exit 1
+	fi
+
+	if [[ $OSTYPE != "msys" ]] ; then
+		cmake --install . --prefix $dir_install --config Release
+		rm -rf Release
+		mv $dir_install/$1 Release
+		rm -rf $dir_install
+	fi
 }
 
 build_vm()
 {
-	rm -rf ruc-vm
-	git clone -b $vm_release --recursive https://github.com/andrey-terekhov/RuC-VM ruc-vm
+	if ! [[ -z $remove ]] ; then
+		rm -rf ruc-vm
+	fi
 
-	cd ruc-vm
-	mkdir -p build && cd build && cmake ..
+	if ! [[ -d ruc-vm ]] ; then
+		git clone -b $vm_release --recursive https://github.com/andrey-terekhov/RuC-VM ruc-vm
+		cd ruc-vm
+	else
+		cd ruc-vm
+		git checkout $vm_release
+	fi
+
+	if [[ $OSTYPE != "msys" ]] ; then
+		CMAKE_BUILD_TYPE=-DCMAKE_BUILD_TYPE=Release
+	fi
+
+	mkdir -p build && cd build && cmake .. $CMAKE_BUILD_TYPE -DTESTING_EXIT_CODE=$exit_code
 	if ! cmake --build . --config Release ; then
 		exit 1
 	fi
 
 	cd ../..
 	if [[ $OSTYPE == "msys" ]] ; then
-		ruc_interpreter=./ruc-vm/build/Release/ruc-vm
+		interpreter=./ruc-vm/build/Release/ruc-vm
 	else
-		ruc_interpreter=./ruc-vm/build/ruc-vm
+		interpreter=./ruc-vm/build/ruc-vm
 	fi
+
+	interpreter_debug=$interpreter
 }
 
 build()
 {
 	cd `dirname $0`/..
-	mkdir -p build && cd build && cmake ..
-	if ! cmake --build . --config Release ; then
-		exit 1
-	fi
+	build_folder ruc
 
-	if [[ $OSTYPE == "msys" ]] ; then
-		ruc_compiler=./Release/ruc.exe
+	compiler=./Release/ruc
+	if [[ -z $fast ]] ; then
+		compiler_debug=./Debug/ruc
 	else
-		ruc_compiler=./ruc
+		compiler_debug=$compiler
 	fi
 
 	if [[ -z $ignore ]] ; then
@@ -107,20 +181,51 @@ build()
 	fi
 }
 
-internal_timeout()
+run()
 {
-	if [[ $OSTYPE == "darwin" ]] ; then
-		gtimeout $@
-	else
-		timeout $@
+	exec=$1
+	exec_debug=$2
+	shift
+	shift
+
+	$runner $exec $@ &>$log
+	ret=$?
+
+	build_type=""
+	if [[ $exec != $exec_debug ]] ; then
+		if [[ $ret == 0 ]] ; then
+			mv $vm_exec $buf
+
+			$runner $exec_debug $@ &>$log
+			ret=$?
+
+			if [[ $ret == 0 ]] ; then
+				mv $buf $vm_exec
+			else
+				build_type="(Debug)"
+			fi
+		elif [[ $ret == $exit_code ]] ; then
+			mv $log $buf
+
+			$runner $exec_debug $@ &>$log
+			ret=$?
+
+			if [[ $ret == $exit_code ]] ; then
+				mv $buf $log
+			else
+				build_type="(Debug)"
+			fi
+		fi
 	fi
+
+	return $ret
 }
 
 message_success()
 {
 	if [[ -z $debug ]] ; then
 		if [[ -z $silence ]] ; then
-			echo -e "\x1B[1;32m $action success \x1B[1;39m: $path"
+			echo -e "\x1B[1;32m $action success \x1B[1;39m: $path $build_type"
 			sleep $output_time
 		fi
 	fi
@@ -129,7 +234,7 @@ message_success()
 message_warning()
 {
 	if [[ -z $silence ]] ; then
-		echo -e "\x1B[1;33m $action warning \x1B[1;39m: $path"
+		echo -e "\x1B[1;33m $action warning \x1B[1;39m: $path $build_type"
 		sleep $output_time
 	fi
 }
@@ -137,7 +242,7 @@ message_warning()
 message_timeout()
 {
 	if [[ -z $silence ]] ; then
-		echo -e "\x1B[1;34m $action timeout \x1B[1;39m: $path"
+		echo -e "\x1B[1;34m $action timeout \x1B[1;39m: $path $build_type"
 		sleep $output_time
 	fi
 }
@@ -145,46 +250,49 @@ message_timeout()
 message_failure()
 {
 	if [[ -z $silence ]] ; then
-		echo -e "\x1B[1;31m $action failure \x1B[1;39m: $path"
+		echo -e "\x1B[1;31m $action failure \x1B[1;39m: $path $build_type"
 		sleep $output_time
 	fi
 }
 
 execution()
 {
-	if [[ $path == $exec_dir/* ]] ; then
+	if [[ $path == $dir_exec/* ]] ; then
 		action="execution"
-		internal_timeout $wait_for $ruc_interpreter export.txt &>$log
+		run $interpreter $interpreter_debug $vm_exec
 
-		case "$?" in
+		case $? in
 			0)
-				if [[ $path == */$error_subdir/* ]] ; then
+				if [[ $path == */$subdir_error/* ]] ; then
 					message_failure
 					let failure++
 				else
-					message_success
-					let success++
+					if [[ $build_type == "(Debug)" ]] ; then
+						build_type=""
+
+						message_failure
+						let failure++
+					else
+						message_success
+						let success++
+					fi
 				fi
 				;;
 			124|142)
 				message_timeout
 				let timeout++
 				;;
-			139|134)
-				# Segmentation fault
-				# Double free or corruption (!prev)
+			$exit_code)
+				if [[ $path == */$subdir_error/* ]] ; then
+					if [[ $build_type == "(Debug)" ]] ; then
+						build_type=""
 
-				message_failure
-				let failure++
-
-				if ! [[ -z $debug ]] ; then
-					cat $log
-				fi
-				;;
-			*)
-				if [[ $path == */$error_subdir/* ]] ; then
-					message_success
-					let success++
+						message_failure
+						let failure++
+					else
+						message_success
+						let success++
+					fi
 				else
 					message_failure
 					let failure++
@@ -192,6 +300,18 @@ execution()
 					if ! [[ -z $debug ]] ; then
 						cat $log
 					fi
+				fi
+				;;
+			*)
+				# Segmentation fault
+				# Double free or corruption (!prev)
+				# Etcetera
+
+				message_failure
+				let failure++
+
+				if ! [[ -z $debug ]] ; then
+					cat $log
 				fi
 				;;
 		esac
@@ -202,7 +322,7 @@ execution()
 
 check_warnings()
 {
-	if [[ $path == */$warning_subdir/* ]] ; then
+	if [[ $path == */$subdir_warning/* ]] ; then
 		message_success
 		let success++
 	else
@@ -228,22 +348,29 @@ check_warnings()
 
 compiling()
 {
-	if [[ -z $ignore || $path != $error_dir/* ]] ; then
+	if [[ -z $ignore || $path != $dir_error/* ]] ; then
 		action="compiling"
-		internal_timeout $wait_for $ruc_compiler $sources &>$log
+		run $compiler $compiler_debug $sources -o $vm_exec
 
-		case "$?" in
+		case $? in
 			0)
-				if [[ $path == $error_dir/* ]] ; then
+				if [[ $path == $dir_error/* ]] ; then
 					message_failure
 					let failure++
 				else
-					message_success
+					if [[ $build_type == "(Debug)" ]] ; then
+						build_type=""
 
-					if [[ -z $ignore ]] ; then
-						execution
+						message_failure
+						let failure++
 					else
-						let success++
+						message_success
+
+						if [[ -z $ignore ]] ; then
+							execution
+						else
+							let success++
+						fi
 					fi
 				fi
 				;;
@@ -251,20 +378,16 @@ compiling()
 				message_timeout
 				let timeout++
 				;;
-			139|134)
-				# Segmentation fault
-				# Double free or corruption (!prev)
+			$exit_code)
+				if [[ $path == $dir_error/* ]] ; then
+					if [[ $build_type == "(Debug)" ]] ; then
+						build_type=""
 
-				message_failure
-				let failure++
-
-				if ! [[ -z $debug ]] ; then
-					cat $log
-				fi
-				;;
-			*)
-				if [[ $path == $error_dir/* ]] ; then
-					check_warnings
+						message_failure
+						let failure++
+					else
+						check_warnings
+					fi
 				else
 					message_failure
 					let failure++
@@ -274,6 +397,18 @@ compiling()
 					fi
 				fi
 				;;
+			*)
+				# Segmentation fault
+				# Double free or corruption (!prev)
+				# Etcetera
+
+				message_failure
+				let failure++
+
+				if ! [[ -z $debug ]] ; then
+					cat $log
+				fi
+				;;
 		esac
 	fi
 }
@@ -281,16 +416,16 @@ compiling()
 test()
 {
 	# Do not use names with spaces!
-	for path in `find $test_dir -name *.c`
+	for path in `find $dir_test -name *.c`
 	do
 		sources=$path
 
-		if [[ $path != */$include_subdir/* ]] ; then
+		if [[ $path != */$subdir_include/* ]] ; then
 			compiling
 		fi
 	done
 
-	for include in `find $test_dir -name $include_subdir -type d`
+	for include in `find $dir_test -name $subdir_include -type d`
 	do
 		for path in `ls -d $include/*`
 		do
@@ -311,7 +446,8 @@ test()
 	fi
 
 	echo -e "\x1B[1;39m success = $success, warning = $warning, failure = $failure, timeout = $timeout"
-	rm $log
+	rm -f $log
+	rm -f $buf
 }
 
 main()
@@ -319,7 +455,11 @@ main()
 	init $@
 
 	build
-	test
+	if [[ -z $debug ]] ; then
+		test 2>/dev/null
+	else
+		test
+	fi
 
 	if [[ $failure != 0 || $timeout != 0 ]] ; then
 		exit 1
